fix migrations

+++ database/migrations/2025_06_05_043324_create_person_notice_signature_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('person_notice_signature', function (Blueprint $table) {
             $table->id('row_id')->autoIncrement();
-            $table->bigInteger('notice_id');
-            $table->bigInteger('notice_person_id');
+            $table->integer('notice_id');
+            $table->integer('notice_person_id');
             $table->foreign('notice_id')->references('notice_id')->on('person_notice')->onDelete('cascade');
             $table->timestamps();
             $table->softDeletes();
