+++ database/migrations/2025_06_07_163045_create_person_fraction.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('person_fraction', function (Blueprint $table) {
             $table->id('row_id');
-            $table->bigInteger('person_id');
-            $table->bigInteger('fraction_id');
+            $table->integer('person_id');
+            $table->integer('fraction_id');
             $table->timestamps();
             $table->softDeletes();
         });
