+++ database/migrations/2025_06_07_165600_create_person_commission.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('person_commission', function (Blueprint $table) {
             $table->id('row_id');
-            $table->bigInteger('person_id');
-            $table->bigInteger('commission_id');
+            $table->integer('person_id');
+            $table->integer('commission_id');
             $table->timestamps();
             $table->softDeletes();
         });
